pyproject.toml
[tool.poetry]
name = "template_ml"
version = "0.0.1"
description = "Machine Learning template"
authors = ["Andres Di Giovanni <andresdigiovanni@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pyyaml = "^6.0.2"
scikit-learn = "^1.5.2"
optuna = "^4.0.0"
xgboost = "^2.1.2"
numpy = "1.26.4"
evidently = "^0.5.0"
cross-ml = "^0.2.0"
pandas = "^2.2.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = [".venv", "__init__.py"]
